# src/db/models/Contact.js
import { model, Schema } from "mongoose";

const contactSchema = new Schema(
  {
    name: { type: String, required: true },
    phoneNumber: { type: String, required: true },
    email: { type: String },
    isFavourite: { type: Boolean, default: false },
    contactType: {
      type: String,
      enum: ["work", "home", "personal"],
      default: "personal",
    },
    photo: { type: String },
    owner: { type: Schema.Types.ObjectId, ref: "users" },
  },
  { timestamps: true, versionKey: false },
);

export const Contact = model("contacts", contactSchema);